# Writerside/api/api.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
...
  /api/admin/studentClass:
    get:
      tags:
      - admin-student-class-controller
      description: Getting student class
      operationId: getStudentClass
      parameters:
      - name: studentClassId
        in: query
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Found student class
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentClassResponse'
        "404":
          description: Student class could not be found.
          content:
            application/json:
              schema:
                type: string
    put:
      tags:
      - admin-student-class-controller
      description: Editing student class
      operationId: editStudentClass
      parameters:
      - name: studentClassId
        in: query
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StudentClassCreationRequest'
        required: true
      responses:
        "400":
          description: A student class with this name already exists.
          content:
            application/json:
              schema:
                type: string
        "404":
          description: Teacher could not be found.
          content:
            application/json:
              schema:
                type: string
        "200":
          description: Edited student class
          content:
            '*/*':
              schema:
                type: object
    post:
      tags:
      - admin-student-class-controller
      description: Create a new student class
      operationId: createStudentClass
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StudentClassCreationRequest'
        required: true
      responses:
        "400":
          description: A student class with this name already exists.
          content:
            application/json:
              schema:
                type: string
        "201":
          description: Created new student class
          content:
            '*/*':
              schema:
                type: object
        "404":
          description: Teacher could not be found.
          content:
            application/json:
              schema:
                type: string
    delete:
      tags:
      - admin-student-class-controller
      description: Deleting student class
      operationId: deleteStudentClass
      parameters:
      - name: studentClassId
        in: query
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: Deleted student class
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StudentClassResponse'
        "404":
          description: Student class could not be found.
          content:
            application/json:
              schema:
                type: string
...
  /api/admin/admin:
    get:
      tags:
      - admin-admin-controller
      description: Gets Admin
      operationId: getAdmin
      parameters:
      - name: adminId
        in: query
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Admin could not be found.
          content:
            application/json:
              schema:
                type: string
        "200":
          description: Found Admin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminResponse'
    put:
      tags:
      - admin-admin-controller
      description: Edit Admin
      operationId: editAdmin
      parameters:
      - name: adminId
        in: query
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminSignupRequest'
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: object
        "404":
          description: Admin could not be found.
          content:
            application/json:
              schema:
                type: string
    post:
      tags:
      - admin-admin-controller
      description: Signs up new Admin
      operationId: createAdmin
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminSignupRequest'
        required: true
      responses:
        "201":
          description: Signed up new Admin
          content:
            '*/*':
              schema:
                type: object
    delete:
      tags:
      - admin-admin-controller
      description: Deletes Admin
      operationId: deleteAdmin
      parameters:
      - name: adminId
        in: query
        required: true
        schema:
          type: integer
          format: int64
      responses:
        "404":
          description: Admin could not be found.
          content:
            application/json:
              schema:
                type: string
        "200":
          description: Deleted Admin
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminResponse'
...
  /api/admin/studentClasses:
    get:
      tags:
      - admin-student-class-controller
      description: Getting all student classes
      operationId: getAllStudentClasses
      responses:
        "200":
          description: Found all student classes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StudentClassResponse'
    delete:
      tags:
      - admin-student-class-controller
      description: Deleting list of student classes
      operationId: deleteStudentClasses
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                type: integer
                format: int64
        required: true
      responses:
        "200":
          description: Deleted student classes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StudentClassResponse'
        "404":
          description: Student class could not be found.
          content:
            application/json:
              schema:
                type: string
...
components:
  schemas:
...
    StudentClassCreationRequest:
      type: object
      properties:
        name:
          type: string
        year:
          type: integer
          format: int32
        teacherId:
          type: integer
          format: int64
...
    TeacherResponse:
      type: object
      properties:
        userId:
          type: integer
          format: int64
        username:
          type: string
        firstname:
          type: string
        surname:
          type: string
        generatedPassword:
          type: string
        teacherId:
          type: integer
          format: int64
        abbreviation:
          type: string
    AdminSignupRequest:
      type: object
      properties:
        firstname:
          type: string
        surname:
          type: string
...
    StudentClassResponse:
      type: object
      properties:
        studentClassId:
          type: integer
          format: int64
        name:
          type: string
        students:
          type: array
          items:
            $ref: '#/components/schemas/StudentResponse'
        teacher:
          $ref: '#/components/schemas/TeacherResponse'
        year:
          type: integer
          format: int32
        releaseYear:
          type: integer
          format: int32
    StudentResponse:
      type: object
      properties:
        userId:
          type: integer
          format: int64
        username:
          type: string
        firstname:
          type: string
        surname:
          type: string
        generatedPassword:
          type: string
        studentId:
          type: integer
          format: int64
        studentClassResponse:
          $ref: '#/components/schemas/StudentClassResponse'
...
    AdminResponse:
      type: object
      properties:
        userId:
          type: integer
          format: int64
        username:
          type: string
        firstname:
          type: string
        surname:
          type: string
        generatedPassword:
          type: string
        adminId:
          type: integer
          format: int64
...